requirements.txt
# Django + API
Django>=5.0.2
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
djangorestframework-simplejwt>=5.3.1

# Database
PyMySQL>=1.0.3
cryptography>=3.4.8

# AI / ML
tensorflow>=2.20.0
scikit-learn>=1.6.0
opencv-python>=4.10.0.84
numpy>=1.26.4
pandas>=2.2.1

# Fairness Auditing
aif360>=0.6.1

# Testing
pytest>=8.2.0
pytest-django>=4.8.0
coverage>=7.0.0

# Utilities
python-dotenv>=1.0.1
gunicorn>=22.0.0
Pillow>=10.0.0

# Additional packages for research
seaborn>=0.12.0
plotly>=5.17.0
jupyter>=1.0.0

